Rename ...FFTF -> PadAnd...FFT
