+++ src/main/java/net/imagej/ops/deconvolve/PadAndRichardsonLucyTV.java
@@ -27,72 +27,52 @@
  * #L%
  */
 
-package net.imagej.ops.filter.correlate;
+package net.imagej.ops.deconvolve;
 
-import net.imagej.ops.Contingent;
 import net.imagej.ops.Ops;
-import net.imagej.ops.filter.AbstractFFTFilterF;
-import net.imagej.ops.special.computer.BinaryComputerOp;
 import net.imagej.ops.special.computer.Computers;
+import net.imagej.ops.special.computer.UnaryComputerOp;
 import net.imglib2.RandomAccessibleInterval;
-import net.imglib2.outofbounds.OutOfBoundsMirrorFactory;
-import net.imglib2.outofbounds.OutOfBoundsMirrorFactory.Boundary;
 import net.imglib2.type.NativeType;
 import net.imglib2.type.numeric.ComplexType;
 import net.imglib2.type.numeric.RealType;
-import net.imglib2.util.Intervals;
 
 import org.scijava.Priority;
+import org.scijava.plugin.Parameter;
 import org.scijava.plugin.Plugin;
 
 /**
- * Correlate op for (@link Img)
+ * Richardson Lucy with total variation function op that operates on (@link
+ * RandomAccessibleInterval) (Richardson-Lucy algorithm with total variation
+ * regularization for 3D confocal microscope deconvolution Microsc Res Rech 2006
+ * Apr; 69(4)- 260-6)
  * 
- * @author Brian Northan
+ * @author bnorthan
  * @param <I>
  * @param <O>
  * @param <K>
  * @param <C>
  */
-@Plugin(type = Ops.Filter.Correlate.class, priority = Priority.VERY_HIGH)
-public class CorrelateFFTF<I extends RealType<I> & NativeType<I>, O extends RealType<O> & NativeType<O>, K extends RealType<K> & NativeType<K>, C extends ComplexType<C> & NativeType<C>>
-	extends AbstractFFTFilterF<I, O, K, C> implements Contingent,
-	Ops.Filter.Correlate
+@Plugin(type = Ops.Deconvolve.RichardsonLucyTV.class,
+	priority = Priority.HIGH)
+public class PadAndRichardsonLucyTV<I extends RealType<I> & NativeType<I>, O extends RealType<O> & NativeType<O>, K extends RealType<K> & NativeType<K>, C extends ComplexType<C> & NativeType<C>>
+	extends PadAndRichardsonLucy<I, O, K, C> implements
+	Ops.Deconvolve.RichardsonLucyTV
 {
 
+	@Parameter
+	private float regularizationFactor = 0.01f;
+
 	@Override
-	public void initialize() {
-
-		// TODO: try and figure out if there is a better place to set the default
-		// OBF
-		if (this.getOBFInput() == null) {
-			setOBFInput(new OutOfBoundsMirrorFactory<>(Boundary.SINGLE));
-		}
-
-		super.initialize();
-
-	}
-
-	/**
-	 * create a correlation filter computer
-	 */
-	@Override
-	public
-		BinaryComputerOp<RandomAccessibleInterval<I>, RandomAccessibleInterval<K>, RandomAccessibleInterval<O>>
-		createFilterComputer(RandomAccessibleInterval<I> paddedInput,
-			RandomAccessibleInterval<K> paddedKernel,
-			RandomAccessibleInterval<C> fftImg, RandomAccessibleInterval<C> fftKernel,
-			RandomAccessibleInterval<O> output)
+	@SuppressWarnings({ "unchecked", "rawtypes" })
+	protected
+		UnaryComputerOp<RandomAccessibleInterval<O>, RandomAccessibleInterval<O>>
+		getComputeEstimateOp()
 	{
-		return Computers.binary(ops(), CorrelateFFTC.class, output,
-			paddedInput, paddedKernel, fftImg, fftKernel);
+		// create Richardson Lucy TV update op, this will override the base RL
+		// Update.
+		return ((UnaryComputerOp) Computers.unary(ops(),
+			RichardsonLucyTVUpdate.class, RandomAccessibleInterval.class,
+			RandomAccessibleInterval.class, regularizationFactor));
 	}
-
-	@Override
-	public boolean conforms() {
-		// TODO: only conforms if the kernel is sufficiently large (else the
-		// naive approach should be used) -> what is a good heuristic??
-		return Intervals.numElements(in2()) > 9;
-	}
-
 }
