+++ src/main/java/net/imagej/ops/filter/correlate/PadAndCorrelateFFT.java
@@ -27,24 +27,26 @@
  * #L%
  */
 
-package net.imagej.ops.filter.convolve;
+package net.imagej.ops.filter.correlate;
 
+import net.imagej.ops.Contingent;
 import net.imagej.ops.Ops;
-import net.imagej.ops.filter.AbstractFFTFilterF;
+import net.imagej.ops.filter.AbstractPadAndFFTFilter;
 import net.imagej.ops.special.computer.BinaryComputerOp;
 import net.imagej.ops.special.computer.Computers;
 import net.imglib2.RandomAccessibleInterval;
-import net.imglib2.outofbounds.OutOfBoundsConstantValueFactory;
+import net.imglib2.outofbounds.OutOfBoundsMirrorFactory;
+import net.imglib2.outofbounds.OutOfBoundsMirrorFactory.Boundary;
 import net.imglib2.type.NativeType;
 import net.imglib2.type.numeric.ComplexType;
 import net.imglib2.type.numeric.RealType;
-import net.imglib2.util.Util;
+import net.imglib2.util.Intervals;
 
 import org.scijava.Priority;
 import org.scijava.plugin.Plugin;
 
 /**
- * Convolve op for (@link Img)
+ * Correlate op for (@link Img)
  * 
  * @author Brian Northan
  * @param <I>
@@ -52,19 +54,19 @@ import org.scijava.plugin.Plugin;
  * @param <K>
  * @param <C>
  */
-@Plugin(type = Ops.Filter.Convolve.class, priority = Priority.HIGH)
-public class ConvolveFFTF<I extends RealType<I> & NativeType<I>, O extends RealType<O> & NativeType<O>, K extends RealType<K> & NativeType<K>, C extends ComplexType<C> & NativeType<C>>
-	extends AbstractFFTFilterF<I, O, K, C> implements Ops.Filter.Convolve
+@Plugin(type = Ops.Filter.Correlate.class, priority = Priority.VERY_HIGH)
+public class PadAndCorrelateFFT<I extends RealType<I> & NativeType<I>, O extends RealType<O> & NativeType<O>, K extends RealType<K> & NativeType<K>, C extends ComplexType<C> & NativeType<C>>
+	extends AbstractPadAndFFTFilter<I, O, K, C> implements Contingent,
+	Ops.Filter.Correlate
 {
 
 	@Override
 	public void initialize() {
 
-		// if no OBF was passed in set default OBF for convolve to be constant value
-		// of zero (zero-pad)
+		// TODO: try and figure out if there is a better place to set the default
+		// OBF
 		if (this.getOBFInput() == null) {
-			setOBFInput(new OutOfBoundsConstantValueFactory<>(Util
-				.getTypeFromInterval(in()).createVariable()));
+			setOBFInput(new OutOfBoundsMirrorFactory<>(Boundary.SINGLE));
 		}
 
 		super.initialize();
@@ -72,7 +74,7 @@ public class ConvolveFFTF<I extends RealType<I> & NativeType<I>, O extends RealT
 	}
 
 	/**
-	 * create a convolve filter computer
+	 * create a correlation filter computer
 	 */
 	@Override
 	public
@@ -82,8 +84,15 @@ public class ConvolveFFTF<I extends RealType<I> & NativeType<I>, O extends RealT
 			RandomAccessibleInterval<C> fftImg, RandomAccessibleInterval<C> fftKernel,
 			RandomAccessibleInterval<O> output)
 	{
-		return Computers.binary(ops(), ConvolveFFTC.class, output, paddedInput,
-			paddedKernel, fftImg, fftKernel);
+		return Computers.binary(ops(), CorrelateFFTC.class, output,
+			paddedInput, paddedKernel, fftImg, fftKernel);
+	}
+
+	@Override
+	public boolean conforms() {
+		// TODO: only conforms if the kernel is sufficiently large (else the
+		// naive approach should be used) -> what is a good heuristic??
+		return Intervals.numElements(in2()) > 9;
 	}
 
 }
